src/api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
black==23.3.0
ruff==0.0.272
alembic==1.12.0
python-dotenv==1.0.0
fastapi-sqlalchemy==0.2.1
sqlalchemy-utils==0.41.1
psycopg2==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4